Fix setup script and rename to install (tests not updated).
