+++ requirements.txt
@@ -1,9 +1,10 @@
-mock==4.0.2
-coverage==5.1
-pytest==5.4.2
-pytest-cov==2.9.0
-coveralls==3.1.0
-mypy==0.770
-mypy-extensions==0.4.3
-flake8==3.8.2
-pyflakes==2.2.0
+mock
+coverage
+pytest
+pytest-cov
+coveralls
+mypy
+mypy-extensions
+flake8
+pyflakes
+typed-ast
